pyproject.toml
[project]
name = "gestionfondos-mobile"
version = "0.1.0"
description = "GestionFondos Mobile App - Personal Finance Manager (Android/iOS)"
requires-python = ">=3.10"
readme = "README.md"
authors = [{name = "GestionFondos Team"}]
dependencies = [
    "Kivy>=2.2.1",
    "KivyMD>=1.2.0",
    "SQLAlchemy>=2.0.0",
    "aiohttp>=3.9.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "python-dateutil>=2.8.2",
    "cryptography>=41.0.0",
    "keyring>=24.0.0",
    "asyncio-contextmanager>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest >= 7.4.0",
    "pytest-asyncio >= 0.21.0",
    "pytest-cov >= 4.1.0",
    "black >= 23.0.0",
    "isort >= 5.12.0",
    "mypy >= 1.0.0",
    "ruff >= 0.1.0",
]
mobile = [
    "buildozer >= 1.5.0",
    "cython >= 0.29.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
strict = false
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
